Relax database check in Request
This changes relaxes the checking we do in Request to the bare minimum -
does the file exist in the index or not. If it does, we don't care about
bits or sizes and just try the read directly. This cuts the time to sync
50k files locally to a third (compare wall time or sender utime):

	Before:

	jb@syno:~/src/github.com/syncthing/syncthing/test (master) $ go test -v -tags "integration benchmark" -run TransferMany
	=== RUN TestBenchmarkTransferManyFiles
	...
	2015/07/11 16:27:19 Verifying...
	2015/07/11 16:27:22 Result: Wall time: 1m42.899254903s
	2015/07/11 16:27:22 Result: 1.5 MiB/s synced
	2015/07/11 16:27:22 Receiver: Utime: 1m14.600862s
	2015/07/11 16:27:22 Receiver: Stime: 32.538708s
	2015/07/11 16:27:22 Receiver: MaxRSS: 160968 KiB
	2015/07/11 16:27:22 Sender: Utime: 3m6.626108s
	2015/07/11 16:27:22 Sender: Stime: 17.583095s
	2015/07/11 16:27:22 Sender: MaxRSS: 135736 KiB
	--- PASS: TestBenchmarkTransferManyFiles (150.20s)

	After:

	jb@syno:~/src/github.com/syncthing/syncthing/test (master) $ go test -v -tags "integration benchmark" -run TransferMany
	=== RUN TestBenchmarkTransferManyFiles
	...
	2015/07/11 16:29:13 Verifying...
	2015/07/11 16:29:15 Result: Wall time: 33.229851726s
	2015/07/11 16:29:15 Result: 4.6 MiB/s synced
	2015/07/11 16:29:15 Receiver: Utime: 50.587266s
	2015/07/11 16:29:15 Receiver: Stime: 26.204354s
	2015/07/11 16:29:15 Receiver: MaxRSS: 138968 KiB
	2015/07/11 16:29:15 Sender: Utime: 49.944314s
	2015/07/11 16:29:15 Sender: Stime: 10.320827s
	2015/07/11 16:29:15 Sender: MaxRSS: 125692 KiB
	--- PASS: TestBenchmarkTransferManyFiles (81.69s)

This will allow requests to files that exists but are flagged as
invalid, which we didn't allow before. I don't see that as a big deal...
The other actually invalid cases are still caught by the exists check or
the actual read (as proven by TestRequest).

We *do* need to do the exists check and not (for example) just check
that the name is within the folder, to prevent requests following
symlinks when they should not and things like that.