lib/db: Disable database compression on slow CPUs
Compression of database blocks obviously has a cost - usually this is
offset by having to shuffle fewer blocks, plus it's nice to have smaller
files on disk. On slow CPUs however, things move along a bit faster if
it doesn't have to do compression as well. This introduces a new config
option that is a threshold on the measured hash rate (as that is our CPU
benchmark). By default, if the hash rate is less than 25 MB/s we disable
database compression. On a Raspberry Pi 2, disabling compression has the
following effect on the database benchmarks:

	benchmark                         old ns/op       new ns/op       delta
	BenchmarkReplaceAll-4             25975055174     28018666889     +7.87%
	BenchmarkUpdateOneChanged-4       1035281         1047867         +1.22%
	BenchmarkUpdateOneUnchanged-4     1016674         1007290         -0.92%
	BenchmarkNeedHalf-4               976915147       813343682       -16.74%
	BenchmarkHave-4                   162623279       160097015       -1.55%
	BenchmarkGlobal-4                 1504958530      1290473498      -14.25%
	BenchmarkNeedHalfTruncated-4      896829925       740321966       -17.45%
	BenchmarkHaveTruncated-4          120785617       122616107       +1.52%
	BenchmarkGlobalTruncated-4        1421247427      1139823799      -19.80%
	BenchmarkUpdateOneFile-4          2076234         2059338         -0.81%

(No difference on allocations.)

Mostly it makes the full table scans a lot faster, and we still do those quite
often.